Add Acme patch for expanding tabs to spaces
Labelled as experimental since I just wrote it and have only put it
through basic tests.

I'm sure there are other patches or Acme forks which implement the
same thing.  However, this implementation is my own, not based on any
prior work, with one minor exception.  The choice of "Spaces" for the
command name was inspired by the gist linked below.  I assume there
is some fork or patch out there which the gist depends on, but I never
found it.

https://gist.github.com/mkhl/5e4cda4f9a262f432eacd592aba5fd54