Make patch license simpler
On second thought, it's confusing to say that my original portions of
the patch are 0BSD, since there's so much ambiguity about what should be
considered original.  Some of the code was modified or duplicated and
modified: is that original?

Let's just say that acme-tab-expand.patch is under the same MIT license
as the rest of Acme.  The puts the whole patch, the resulting patched
work, under the same license.  Keep it simple.