9a: add -i and -c ncol options
Add -i option to start with Spaces ON.  Same as the -i option added to
Acme by spew's Spaces patch.

Add -c ncol option to control the initial number of columns.  Same as
the -c ncol option in Acme.  However, change the default number of
columns from 2 to 1, so that Acme doesn't start with an empty column
that has a blindingly white background.