9a/9s: tolerate $PLAN9 being unset if 9 is in PATH
Tolerate the $PLAN9 variable being unset, provided that the 9 script is
in the path, which can be used to set $PLAN9 (and add $PLAN9/bin to the
PATH).

Doing this for NixOS: its plan9port package automatically puts 9 in
the PATH, but not other executables like acme or sam, and the PLAN9
variable cannot be pointed at a static directory, since its location
in the Nix store changes.